+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added historical changelog to GIT repo
 - In context service provider, renamed registerFeatures to registerContexts
+- Added ContextLoading and ContextLoaded events
 
 ## [2.2.1 (2019-03-11)](https://github.com/Morning-Train/LaravelContext/compare/2.2.0...2.2.1)